+++ ubiquity_template/library/r_general/ubiquity.r
@@ -26,7 +26,7 @@
 #'@importFrom officer ph_location_label ph_with read_pptx read_docx shortcuts  slip_in_seqfield slip_in_text 
 #'@importFrom officer styles_info unordered_list
 #'@importFrom PKNCA PKNCA.options PKNCAconc PKNCAdose PKNCAdata pk.nca get.interval.cols
-#'@importFrom utils read.csv read.delim txtProgressBar setTxtProgressBar write.csv tail packageVersion sessionInfo
+#'@importFrom utils capture.output read.csv read.delim txtProgressBar setTxtProgressBar write.csv tail packageVersion sessionInfo
 #'@importFrom stats median qt var sd
 #'@importFrom MASS mvrnorm
 
@@ -6403,6 +6403,7 @@ odtest = calculate_objective(cfg$estimation$parameters$guess, cfg, estimation=FA
       }
     message(pstr)
     pest$sysup = paste(pest$sysup, pstr, "\n")
+    }
 
     # Writing system update text to a file
     sysup_file =file.path(cfg[["options"]][["misc"]][["output_directory"]], "system_update.txt")
@@ -6412,10 +6413,12 @@ odtest = calculate_objective(cfg$estimation$parameters$guess, cfg, estimation=FA
 
     # Writing session information to a file
     SI_file = file.path(cfg[["options"]][["misc"]][["output_directory"]], "sessionInfo.RData")
+    SI_text = file.path(cfg[["options"]][["misc"]][["output_directory"]], "sessionInfo.txt")
     SI = sessionInfo()
+    # SI object
     save(SI, file=SI_file)
-
-    }
+    # SI text
+    utils::capture.output(sessionInfo(), file=SI_text)
 
   } else {
     vp(cfg, sprintf('The estimation was terminated. We were unable to   '))
@@ -7144,12 +7147,14 @@ f.source      = c(f.source,      file.path(output_directory, "parameters_all.csv
 f.source      = c(f.source,      file.path(output_directory, "parameters_est.csv"))
 f.source      = c(f.source,      file.path(output_directory, "report.txt"        ))
 f.source      = c(f.source,      file.path(output_directory, "sessionInfo.RData" ))
+f.source      = c(f.source,      file.path(output_directory, "sessionInfo.txt" ))
 f.source      = c(f.source,      file.path(output_directory, "system_update.txt" ))
 
 f.destination = c(f.destination, file.path(output_directory, paste(name, "-parameters_all.csv", sep="")))
 f.destination = c(f.destination, file.path(output_directory, paste(name, "-parameters_est.csv", sep="")))
 f.destination = c(f.destination, file.path(output_directory, paste(name, "-report.txt"        , sep="")))
 f.destination = c(f.destination, file.path(output_directory, paste(name, "-sessionInfo.RData" , sep="")))
+f.destination = c(f.destination, file.path(output_directory, paste(name, "-sessionInfo.txt" ,   sep="")))
 f.destination = c(f.destination, file.path(output_directory, paste(name, "-system_update.txt" , sep="")))
 
 # clearing out the destination files to prevent old results from lingering
@@ -9498,11 +9503,11 @@ system_report_slide_content = function (cfg,
                                content                = 'Text'){
   #Checking user input:
   isgood = TRUE
-  if(cfg$reporting$enabled){
-    if(rptname %in% names(cfg$reporting$reports)){
-      if( "PowerPoint" != cfg$reporting$reports[[rptname]]$rpttype){
+  if(cfg[["reporting"]][["enabled"]]){
+    if(rptname %in% names(cfg[["reporting"]][["reports"]])){
+      if( "PowerPoint" != cfg[["reporting"]][["reports"]][[rptname]][["rpttype"]]){
         isgood = FALSE
-        vp(cfg, paste("Error: Trying to add PowerPoint content to >", cfg$reporting$reports[[rptname]]$rpttype,"< report", sep=""))
+        vp(cfg, paste("Error: Trying to add PowerPoint content to >", cfg[["reporting"]][["reports"]][[rptname]][["rpttype"]],"< report", sep=""))
       }
     } else {
       isgood = FALSE
@@ -9515,29 +9520,27 @@ system_report_slide_content = function (cfg,
 
   if(isgood){
     # Pulling out the meta data for the report template
-    meta = cfg$reporting$reports[[rptname]]$meta 
-
+    meta = cfg[["reporting"]][["reports"]][[rptname]][["meta"]]
     # Pulling out the report to make it easier to deal with
-    tmprpt  = cfg$reporting$reports[[rptname]]$report
-
+    tmprpt  = cfg[["reporting"]][["reports"]][[rptname]][["report"]]
     # Adding the slide
     if(content_type %in% c("text", "imagefile", "ggplot", "table", "flextable", "flextable_object")){
       tmprpt = officer::add_slide(x      = tmprpt, 
-                         layout = meta$content$layout$general,
-                         master = meta$content$master$general)
-      body_index          = meta$content$indices$content_body
-      sub_title_index     = meta$content$indices$content_sub_title
-      body_ph_label       = meta$content$ph_labels$content_body
-      sub_title_ph_label  = meta$content$ph_labels$content_sub_title
+                         layout = meta[["content"]][["layout"]][["general"]],
+                         master = meta[["content"]][["master"]][["general"]])
+      body_index          = meta[["content"]][["indices"]][["content_body"]]
+      sub_title_index     = meta[["content"]][["indices"]][["content_sub_title"]]
+      body_ph_label       = meta[["content"]][["ph_labels"]][["content_body"]]
+      sub_title_ph_label  = meta[["content"]][["ph_labels"]][["content_sub_title"]]
     }
     else if(content_type == "list"){
       tmprpt = officer::add_slide(x      = tmprpt, 
-                         layout = meta$content$layout$list,
-                         master = meta$content$master$list)
-      body_index          = meta$content$indices$list_body
-      sub_title_index     = meta$content$indices$list_sub_title
-      body_ph_label       = meta$content$ph_labels$list_body
-      sub_title_ph_label  = meta$content$ph_labels$list_sub_title
+                         layout = meta[["content"]][["layout"]][["list"]],
+                         master = meta[["content"]][["master"]][["list"]])
+      body_index          = meta[["content"]][["indices"]][["list_body"]]
+      sub_title_index     = meta[["content"]][["indices"]][["list_sub_title"]]
+      body_ph_label       = meta[["content"]][["ph_labels"]][["list_body"]]
+      sub_title_ph_label  = meta[["content"]][["ph_labels"]][["list_sub_title"]]
     }
 
     # Adding Slide title/subtitle information
@@ -9557,9 +9560,9 @@ system_report_slide_content = function (cfg,
                                       ph_label     = body_ph_label)
   
     # Putting the report back into cfg
-    cfg$reporting$reports[[rptname]]$report = tmprpt
+    cfg[["reporting"]][["reports"]][[rptname]][["report"]] = tmprpt
   } else {
-    vp(cfg, "system_report_slide_content() ")
+    vp(cfg, "system_report_slide_content()")
     vp(cfg, "Unable to add slide, see above for details")
   }
   
@@ -9607,11 +9610,11 @@ system_report_slide_two_col = function (cfg,
 
   #Checking user input:
   isgood = TRUE
-  if(cfg$reporting$enabled){
-    if(rptname %in% names(cfg$reporting$reports)){
-      if( "PowerPoint" != cfg$reporting$reports[[rptname]]$rpttype){
+  if(cfg[["reporting"]][["enabled"]]){
+    if(rptname %in% names(cfg[["reporting"]][["reports"]])){
+      if( "PowerPoint" != cfg[["reporting"]][["reports"]][[rptname]][["rpttype"]]){
         isgood = FALSE
-        vp(cfg, paste("Error: Trying to add PowerPoint content to >", cfg$reporting$reports[[rptname]]$rpttype,"< report", sep=""))
+        vp(cfg, paste("Error: Trying to add PowerPoint content to >", cfg[["reporting"]][["reports"]][[rptname]][["rpttype"]],"< report", sep=""))
       }
     } else {
       isgood = FALSE
@@ -9625,69 +9628,69 @@ system_report_slide_two_col = function (cfg,
 
   if(isgood){
     # Pulling out the meta data for the report template
-    meta = cfg$reporting$reports[[rptname]]$meta 
+    meta = cfg[["reporting"]][["reports"]][[rptname]][["meta"]]
     # Pulling out the report to make it easier to deal with
-    tmprpt  = cfg$reporting$reports[[rptname]]$report
+    tmprpt  = cfg[["reporting"]][["reports"]][[rptname]][["report"]]
 
     #-------------------------------------------------------
     #  here we initialize the correct slide and 
-    #  define the indices
+    #  define the indices/placeholders
     #
     if(content_type %in% c('text')){
       if(is.null(left_content_header) & is.null(right_content_header)){
         # Text without headers
         tmprpt = officer::add_slide(x      = tmprpt, 
-                           layout = meta$two_col$layout$text,
-                           master = meta$two_col$master$text)
+                           layout = meta[["two_col"]][["layout"]][["text"]],
+                           master = meta[["two_col"]][["master"]][["text"]])
 
-        left_index               = meta$two_col$indices$text_left
-        right_index              = meta$two_col$indices$text_right
+        left_index               = meta[["two_col"]][["indices"]][["text_left"]]
+        right_index              = meta[["two_col"]][["indices"]][["text_right"]]
         left_title_index         = NULL
         right_title_index        = NULL
 
-        sub_title_index          = meta$two_col$indices$text_sub_title
-        sub_title_ph_label       = meta$two_col$ph_labels$text_sub_title
-        left_ph_label            = meta$two_col$ph_labels$text_left
-        right_ph_label           = meta$two_col$ph_labels$text_right
+        sub_title_index          = meta[["two_col"]][["indices"]][["text_sub_title"]]
+        sub_title_ph_label       = meta[["two_col"]][["ph_labels"]][["text_sub_title"]]
+        left_ph_label            = meta[["two_col"]][["ph_labels"]][["text_left"]]
+        right_ph_label           = meta[["two_col"]][["ph_labels"]][["text_right"]]
         left_title_ph_label      = NULL
         right_title_ph_label     = NULL
 
       } else {
         # Text with headers
         tmprpt = officer::add_slide(x      = tmprpt, 
-                           layout = meta$two_col$layout$text_head,
-                           master = meta$two_col$master$text_head)
+                           layout = meta[["two_col"]][["layout"]][["text_head"]],
+                           master = meta[["two_col"]][["master"]][["text_head"]])
 
-        left_index               = meta$two_col$indices$text_head_left
-        right_index              = meta$two_col$indices$text_head_right
-        left_title_index         = meta$two_col$indices$text_head_left_title
-        right_title_index        = meta$two_col$indices$text_head_right_title
-        sub_title_index          = meta$two_col$indices$text_head_sub_title
+        left_index               = meta[["two_col"]][["indices"]][["text_head_left"]]
+        right_index              = meta[["two_col"]][["indices"]][["text_head_right"]]
+        left_title_index         = meta[["two_col"]][["indices"]][["text_head_left_title"]]
+        right_title_index        = meta[["two_col"]][["indices"]][["text_head_right_title"]]
+        sub_title_index          = meta[["two_col"]][["indices"]][["text_head_sub_title"]]
+       #sub_title_index          = meta[["two_col"]][["indices"]][["text_head_sub_title"]]
 
-        sub_title_index          = meta$two_col$indices$text_head_sub_title
-        sub_title_ph_label       = meta$two_col$ph_labels$text_head_sub_title
-        left_ph_label            = meta$two_col$ph_labels$text_head_left
-        right_ph_label           = meta$two_col$ph_labels$text_head_right
-        left_title_ph_label      = meta$two_col$ph_labels$text_head_left_title
-        right_title_ph_label     = meta$two_col$ph_labels$text_head_right_title
+        sub_title_ph_label       = meta[["two_col"]][["ph_labels"]][["text_head_sub_title"]]
+        left_ph_label            = meta[["two_col"]][["ph_labels"]][["text_head_left"]]
+        right_ph_label           = meta[["two_col"]][["ph_labels"]][["text_head_right"]]
+        left_title_ph_label      = meta[["two_col"]][["ph_labels"]][["text_head_left_title"]]
+        right_title_ph_label     = meta[["two_col"]][["ph_labels"]][["text_head_right_title"]]
       }
     }else if(content_type %in% c('list')){
       if(is.null(left_content_header) & is.null(right_content_header)){
         # List without headers
         tmprpt = officer::add_slide(x      = tmprpt, 
-                           layout = meta$two_col$layout$list,
-                           master = meta$two_col$master$list)
+                           layout = meta[["two_col"]][["layout"]][["list"]],
+                           master = meta[["two_col"]][["master"]][["list"]])
 
-        left_index               = meta$two_col$indices$list_left
-        right_index              = meta$two_col$indices$list_right
+        left_index               = meta[["two_col"]][["indices"]][["list_left"]]
+        right_index              = meta[["two_col"]][["indices"]][["list_right"]]
         left_title_index         = NULL
         right_title_index        = NULL
-        sub_title_index          = meta$two_col$indices$list_sub_title
+        sub_title_index          = meta[["two_col"]][["indices"]][["list_sub_title"]]
 
-        sub_title_index          = meta$two_col$indices$list_sub_title
-        sub_title_ph_label       = meta$two_col$ph_labels$list_sub_title
-        left_ph_label            = meta$two_col$ph_labels$list_left
-        right_ph_label           = meta$two_col$ph_labels$list_right
+      # sub_title_index          = meta[["two_col"]][["indices"]][["list_sub_title"]]
+        sub_title_ph_label       = meta[["two_col"]][["ph_labels"]][["list_sub_title"]]
+        left_ph_label            = meta[["two_col"]][["ph_labels"]][["list_left"]]
+        right_ph_label           = meta[["two_col"]][["ph_labels"]][["list_right"]]
         left_title_ph_label      = NULL
         right_title_ph_label     = NULL
 
@@ -9696,21 +9699,21 @@ system_report_slide_two_col = function (cfg,
       } else {
         # List with headers
         tmprpt = officer::add_slide(x      = tmprpt, 
-                           layout = meta$two_col$layout$list_head,
-                           master = meta$two_col$master$list_head)
+                           layout = meta[["two_col"]][["layout"]][["list_head"]],
+                           master = meta[["two_col"]][["master"]][["list_head"]])
 
-        left_index         = meta$two_col$indices$list_head_left
-        right_index        = meta$two_col$indices$list_head_right
-        left_title_index   = meta$two_col$indices$list_head_left_title
-        right_title_index  = meta$two_col$indices$list_head_right_title
-        sub_title_index    = meta$two_col$indices$list_head_sub_title
+        left_index         = meta[["two_col"]][["indices"]][["list_head_left"]]
+        right_index        = meta[["two_col"]][["indices"]][["list_head_right"]]
+        left_title_index   = meta[["two_col"]][["indices"]][["list_head_left_title"]]
+        right_title_index  = meta[["two_col"]][["indices"]][["list_head_right_title"]]
+        sub_title_index    = meta[["two_col"]][["indices"]][["list_head_sub_title"]]
 
-        sub_title_index          = meta$two_col$indices$list_head_sub_title
-        sub_title_ph_label       = meta$two_col$ph_labels$list_head_sub_title
-        left_ph_label            = meta$two_col$ph_labels$list_head_left
-        right_ph_label           = meta$two_col$ph_labels$list_head_right
-        left_title_ph_label      = meta$two_col$ph_labels$list_head_left_title
-        right_title_ph_label     = meta$two_col$ph_labels$list_head_right_title
+       #sub_title_index          = meta$two_col$indices$list_head_sub_title
+        sub_title_ph_label       = meta[["two_col"]][["ph_labels"]][["list_head_sub_title"]]
+        left_ph_label            = meta[["two_col"]][["ph_labels"]][["list_head_left"]]
+        right_ph_label           = meta[["two_col"]][["ph_labels"]][["list_head_right"]]
+        left_title_ph_label      = meta[["two_col"]][["ph_labels"]][["list_head_left_title"]]
+        right_title_ph_label     = meta[["two_col"]][["ph_labels"]][["list_head_right_title"]]
       }
     }
 
@@ -9727,8 +9730,6 @@ system_report_slide_two_col = function (cfg,
       tmprpt = officer::ph_with(x=tmprpt, location = officer::ph_location_type(type = "title"),  value=title) } 
     if(!is.null(sub_title_index) & !is.null(sub_title)){
       tmprpt = officer::ph_with(x=tmprpt,  location=officer::ph_location_label(ph_label=sub_title_ph_label), value=sub_title) }
-
-
 
     #
     # Creating the headers
@@ -9776,7 +9777,7 @@ system_report_slide_two_col = function (cfg,
 
 
     # Putting the report back into cfg
-    cfg$reporting$reports[[rptname]]$report = tmprpt
+    cfg[["reporting"]][["reports"]][[rptname]][["report"]] = tmprpt
   } else {
     vp(cfg, "system_report_slide_two_col() ")
     vp(cfg, "Unable to add slide, see above for details")
@@ -9813,11 +9814,11 @@ system_report_slide_section = function (cfg,
 
   #Checking user input:
   isgood = TRUE
-  if(cfg$reporting$enabled){
-    if(rptname %in% names(cfg$reporting$reports)){
-      if( "PowerPoint" != cfg$reporting$reports[[rptname]]$rpttype){
+  if(cfg[["reporting"]][["enabled"]]){
+    if(rptname %in% names(cfg[["reporting"]][["reports"]])){
+      if( "PowerPoint" != cfg[["reporting"]][["reports"]][[rptname]][["rpttype"]]){
         isgood = FALSE
-        vp(cfg, paste("Error: Trying to add PowerPoint content to >", cfg$reporting$reports[[rptname]]$rpttype,"< report", sep=""))
+        vp(cfg, paste("Error: Trying to add PowerPoint content to >", cfg[["reporting"]][["reports"]][[rptname]][["rpttype"]],"< report", sep=""))
       }
     } else {
       isgood = FALSE
@@ -9830,14 +9831,14 @@ system_report_slide_section = function (cfg,
 
   if(isgood){
     # Pulling out the meta data for the report template
-    meta = cfg$reporting$reports[[rptname]]$meta 
+    meta = cfg[["reporting"]][["reports"]][[rptname]][["meta"]]
     # Pulling out the report to make it easier to deal with
-    tmprpt  = cfg$reporting$reports[[rptname]]$report
+    tmprpt  = cfg[["reporting"]][["reports"]][[rptname]][["report"]]
 
     # Adding the title slide
     tmprpt = officer::add_slide(x      = tmprpt, 
-                       layout = meta$section$layout$general,
-                       master = meta$section$master$general)
+                       layout = meta[["section"]][["layout"]][["general"]],
+                       master = meta[["section"]][["master"]][["general"]])
 
 
     # Adding Slide title/subtitle information
@@ -9845,8 +9846,8 @@ system_report_slide_section = function (cfg,
       if(meta$section$type$title == "ctrTitle"){
         tmprpt = officer::ph_with(x=tmprpt,  location = officer::ph_location_type(type = "ctrTitle"), value=title) 
        } else {
-         tmprpt = officer::ph_with(x=tmprpt,  location = officer::ph_location_type(type = meta$section$type$title), 
-                                              index    = meta$section$indices$title, 
+         tmprpt = officer::ph_with(x=tmprpt,  location = officer::ph_location_type(type = meta[["section"]][["type"]][["title"]]), 
+                                              index    = meta[["section"]][["indices"]][["title"]],
                                               value    = title) 
        }
      } 
@@ -9854,14 +9855,14 @@ system_report_slide_section = function (cfg,
       if(meta$section$type$sub_title == "subTitle"){
         tmprpt = officer::ph_with(x=tmprpt,  location = officer::ph_location_type(type = "subTitle"), value=sub_title) 
        } else {
-        tmprpt = officer::ph_with(x=tmprpt,  location = officer::ph_location_type(type = meta$section$type$sub_title), 
-                                    index    = meta$section$indices$sub_title, 
+        tmprpt = officer::ph_with(x=tmprpt,  location = officer::ph_location_type(type = meta[["section"]][["type"]][["sub_title"]]),
+                                    index    = meta[["section"]][["indices"]][["sub_title"]],
                                     value    = sub_title) 
        }
      }
 
     # Putting the report back into cfg
-    cfg$reporting$reports[[rptname]]$report = tmprpt
+    cfg[["reporting"]][["reports"]][[rptname]][["report"]] = tmprpt
   } 
 
   if(!isgood){
@@ -9898,11 +9899,11 @@ system_report_slide_title   = function (cfg,
 
   #Checking user input:
   isgood = TRUE
-  if(cfg$reporting$enabled){
-    if(rptname %in% names(cfg$reporting$reports)){
-      if( "PowerPoint" != cfg$reporting$reports[[rptname]]$rpttype){
+  if(cfg[["reporting"]][["enabled"]]){
+    if(rptname %in% names(cfg[["reporting"]][["reports"]])){
+      if( "PowerPoint" != cfg[["reporting"]][["reports"]][[rptname]][["rpttype"]]){
         isgood = FALSE
-        vp(cfg, paste("Error: Trying to add PowerPoint content to >", cfg$reporting$reports[[rptname]]$rpttype,"< report", sep=""))
+        vp(cfg, paste("Error: Trying to add PowerPoint content to >", cfg[["reporting"]][["reports"]][[rptname]][["rpttype"]],"< report", sep=""))
       }
     } else {
       isgood = FALSE
@@ -9915,37 +9916,36 @@ system_report_slide_title   = function (cfg,
 
   if(isgood){
     # Pulling out the meta data for the report template
-    meta = cfg$reporting$reports[[rptname]]$meta 
+    meta = cfg[["reporting"]][["reports"]][[rptname]][["meta"]]
     # Pulling out the report to make it easier to deal with
-    tmprpt  = cfg$reporting$reports[[rptname]]$report
-
+    tmprpt  = cfg[["reporting"]][["reports"]][[rptname]][["report"]]
     # Adding the title slide
     tmprpt = officer::add_slide(x      = tmprpt, 
-                       layout = meta$title$layout$general,
-                       master = meta$title$master$general)
+                       layout = meta[["title"]][["layout"]][["general"]],
+                       master = meta[["title"]][["master"]][["general"]])
 
     # Adding Slide title/subtitle information
     if(meta$title$type$title == "ctrTitle"){
       tmprpt = officer::ph_with(x=tmprpt,  location = officer::ph_location_type(type = "ctrTitle"), value=title) 
      } else {
-      tmprpt = officer::ph_with(x=tmprpt,  location = officer::ph_location_type(type = meta$title$type$title), 
-                                           index    = meta$title$indices$title, 
+      tmprpt = officer::ph_with(x=tmprpt,  location = officer::ph_location_type(type = meta[["title"]][["type"]][["title"]]), 
+                                           index    = meta[["title"]][["indices"]][["title"]],
                                            value    = title) 
      }
     if(!is.null(sub_title)){
       if(meta$title$type$sub_title == "subTitle"){
         tmprpt = officer::ph_with(x=tmprpt,  location = officer::ph_location_type(type = "subTitle"), value=sub_title) 
        } else {
-        tmprpt = officer::ph_with(x=tmprpt,  location = officer::ph_location_type(type = meta$title$type$sub_title), 
-                                             index    = meta$title$indices$sub_title, 
-                                             value    = sub_title) 
+        tmprpt = officer::ph_with(x=tmprpt,  location = officer::ph_location_type(type = meta[["title"]][["type"]][["sub_title"]]),
+                                    index    = meta[["title"]][["indices"]][["sub_title"]],
+                                    value    = sub_title) 
        }
      }
 
     # Putting the report back into cfg
-    cfg$reporting$reports[[rptname]]$report = tmprpt
+    cfg[["reporting"]][["reports"]][[rptname]][["report"]] = tmprpt
   } else {
-    vp(cfg, "system_report_slide_title() ")
+    vp(cfg, "system_report_slide_title()")
     vp(cfg, "Unable to add slide, see above for details")
 
   }
@@ -11313,17 +11313,19 @@ system_report_estimation = function (cfg,
   if(isgood){
 
     # File names where the estimation results should be stored:
-    fname_estimate = file.path(output_directory, paste(analysis_name, ".RData",    sep=""))
-    fname_grobs    = file.path(output_directory, paste(analysis_name, "_pr.RData", sep=""))
+    fname_estimate = file.path(output_directory, paste(analysis_name, ".RData",          sep=""))
+    fname_grobs    = file.path(output_directory, paste(analysis_name, "_pr.RData",       sep=""))
+    fname_SI_text  = file.path(output_directory, paste(analysis_name, "-sessionInfo.txt", sep=""))
     vp(cfg, "")
     vp(cfg, paste("Appending estimation results to report"))
     vp(cfg, paste("  Report:   ", rptname,            sep=""))
     vp(cfg, paste("  Type:     ", rpttype,            sep=""))
     vp(cfg, paste("  Analysis: ", analysis_name,      sep=""))
     #---------------------------
-    pe    = NULL
-    pest  = NULL
-    grobs = NULL
+    pe      = NULL
+    pest    = NULL
+    grobs   = NULL
+    SI_text = NULL
     if(file.exists(fname_estimate)){
       vp(cfg, paste("Loading estimation results from file:", fname_estimate))
       # Loads the variable pe and pest
@@ -11337,6 +11339,12 @@ system_report_estimation = function (cfg,
       load(fname_grobs)
     } else {
       vp(cfg, paste("Unable to load the figures from file:", fname_grobs))
+    }
+    if(file.exists(fname_SI_text)){
+      vp(cfg, paste("Loading the session information from file:", fname_SI_text))
+      SI_text = readLines(fname_SI_text)
+    } else {
+      vp(cfg, paste("Unable to load the session information from file:", fname_SI_text))
     }
 
     #---------------------------
@@ -11512,6 +11520,19 @@ system_report_estimation = function (cfg,
           content_type  = "text",
           content       = list(style   = "code",
                                text    = line))
+      }
+      # Appending the sessionInfo()
+      if(!is.null(SI_text)){
+        cfg = system_report_doc_add_content(cfg, 
+          content_type  = "text",
+          content       = list(style   = "h2",
+                               text    = "sessionInfo()"))
+         for(line in SI_text){
+           cfg = system_report_doc_add_content(cfg, 
+             content_type  = "text",
+             content       = list(style   = "code",
+                                  text    = line))
+         }
       }
     }
     #---------------------------